requirements.txt
aniso8601==2.0.0
appdirs==1.4.3
astroid==1.6.1
attrs==17.4.0
autopep8==1.3.4
Babel==2.5.3
backports.functools-lru-cache==1.5
certifi==2018.1.18
chardet==3.0.4
click==6.7
configparser==3.5.0
coverage==4.5
cycler==0.10.0
decorator==4.2.1
dummyrdd==0.1.2
enum34==1.1.6
eventlet==0.22.0
Faker==0.8.10
funcsigs==1.0.2
functools32==3.2.3.post2
funcy==1.10
future==0.16.0
futures==3.2.0
gevent==1.2.2
greenlet==0.4.13
idna==2.6
ipaddress==1.0.19
isort==4.3.3
itsdangerous==0.24
Jinja2==2.10
joblib==0.11
lazy-object-proxy==1.3.1
lifelines==0.13.0
MarkupSafe==1.0
marshmallow==2.15.0
matplotlib==2.1.2
mccabe==0.6.1
mock==2.0.0
mockredispy==2.9.3
networkx==1.9
numpy==1.14.0
packaging==16.8
pandas==0.23.4
pbr==3.1.1
pep8==1.7.1
pluggy==0.6.0
ply==3.10
py==1.5.2
py4j==0.10.7
pyaml==17.12.1
pycodestyle==2.3.1
pyinotify==0.9.6
pylint==1.8.2
pyparsing==2.2.0
Pyqtree==0.24
pyshp==1.2.12
pyspark==2.3.0
pytest==3.4.0
pytest-cov==2.5.1
python-dateutil==2.6.1
python-engineio==2.0.2
python-socketio==1.8.4
pytz==2017.3
PyYAML==3.12
redis==2.10.6
requests==2.18.4
RestrictedPython==4.0a1
rq==0.10.0
scikit-learn==0.19.1
scipy==1.0.0
seaborn==0.8.1
simplejson==3.15.0
singledispatch==3.4.0.3
six==1.11.0
sklearn==0.0
subprocess32==3.2.7
text-unidecode==1.1
urllib3==1.22
wrapt==1.10.11
